remote 모듈에 네트워크 관련 코드 옮김
