+++ data/build.gradle.kts
@@ -9,7 +9,6 @@ repositories {
 
 dependencies {
     implementation(project(":domain"))
-    implementation(project(":remote"))
 
     implementation("com.squareup.okhttp3:okhttp:4.10.0")
     implementation("com.squareup.okhttp3:logging-interceptor:4.10.0")
