+++ data/src/main/java/camp/nextstep/edu/github/data/GitRepo.kt
@@ -1,12 +1,6 @@
 package camp.nextstep.edu.github.data
 
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-internal data class GitRepo(
-    @Json(name = "full_name")
+data class GitRepo(
     val fullName: String,
-    @Json(name = "description")
-    val description: String?
+    val description: String?,
 )
