+++ data/src/main/java/camp/nextstep/edu/github/data/GitRepoRepository.kt
@@ -1,4 +1,4 @@
-package camp.nextstep.edu.github.domain
+package camp.nextstep.edu.github.data
 
 interface GitRepoRepository {
     suspend fun getGitRepos(): List<GitRepo>
