--- domain/src/main/java/camp/nextstep/edu/github/domain/GitRepo.kt
@@ -1,6 +1,0 @@
-package camp.nextstep.edu.github.domain
-
-data class GitRepo(
-    val fullName: String,
-    val description: String?,
-)
