+++ remote/src/main/java/camp/nextstep/edu/github/remote/GitRepo.kt
@@ -1,0 +1,12 @@
+package camp.nextstep.edu.github.remote
+
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+internal data class GitRepo(
+    @Json(name = "full_name")
+    val fullName: String,
+    @Json(name = "description")
+    val description: String?
+)
